Create FirefoxESR.download.recipe
I tested this recipe a couple of different ways because the original that I found had some issues (has URLs for both the regular channel release and the ESR release). Also has inputs and a process for MozillaURLProvider, which is a separate script that needs to reside in the same directory as this recipe, even though the link for the latest ESR version is hard-coded.  This is the version that I settled on (hard-coded url) with comments on the code that I will be removing.