Adding another criterion to the template
Adding another criterion that determines whether Office 2016 or Office 2019 is installed on the client machine. Since Microsoft didn't chnage the versioning from 16.x there isn't an easy way to distinguish the two suites. I wrote an extension attribute that makes that determination based on whether the version number of any given app is less than 16.17 or not, and spits out "Office 2016" or "Office 2019". The new criterion only adds a machine to the smart group if 2019 is not reported.